# frontend-integration-package/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowLedger API",
    "version": "0.1.0",
    "description": "API for managing clients, audits, industries, and related business processes"
  },
...
  "paths": {
...
    "/api/contact-social-profiles": {
      "get": {
        "summary": "List contact social profiles",
        "tags": [
          "ContactSocialProfiles"
        ],
        "responses": {
          "200": {
            "description": "Profiles list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "contact_id": {
                            "type": "integer"
                          },
                          "provider": {
                            "type": "string"
                          },
                          "profile_url": {
                            "type": "string"
                          },
                          "is_primary": {
                            "type": "boolean"
                          },
                          "created_utc": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_utc": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create contact social profile",
        "tags": [
          "ContactSocialProfiles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "contact_id",
                  "provider",
                  "profile_url"
                ],
                "properties": {
                  "contact_id": {
                    "type": "integer"
                  },
                  "provider": {
                    "type": "string"
                  },
                  "profile_url": {
                    "type": "string"
                  },
                  "is_primary": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "contact_id": {
                          "type": "integer"
                        },
                        "provider": {
                          "type": "string"
                        },
                        "profile_url": {
                          "type": "string"
                        },
                        "is_primary": {
                          "type": "boolean"
                        },
                        "created_utc": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated_utc": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/contact-social-profiles/{id}": {
      "get": {
        "summary": "Get contact social profile by id",
        "tags": [
          "ContactSocialProfiles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "contact_id": {
                          "type": "integer"
                        },
                        "provider": {
                          "type": "string"
                        },
                        "profile_url": {
                          "type": "string"
                        },
                        "is_primary": {
                          "type": "boolean"
                        },
                        "created_utc": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated_utc": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update contact social profile",
        "tags": [
          "ContactSocialProfiles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contact_id": {
                    "type": "integer"
                  },
                  "provider": {
                    "type": "string"
                  },
                  "profile_url": {
                    "type": "string"
                  },
                  "is_primary": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "contact_id": {
                          "type": "integer"
                        },
                        "provider": {
                          "type": "string"
                        },
                        "profile_url": {
                          "type": "string"
                        },
                        "is_primary": {
                          "type": "boolean"
                        },
                        "created_utc": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated_utc": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete contact social profile",
        "tags": [
          "ContactSocialProfiles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deleted": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}